GSON+Foundation extensions
