+++ TestApp/TestApp/Deepstream Extensions/JavaUtilsArrayList+Extension.swift
@@ -10,12 +10,22 @@ import Foundation
 
 extension JavaUtilArrayList {
     
-    func toNSArray<T>() -> [T] {
+    func toArray<T>() -> [T] {
         var arr:[T] = []
         for i : Int32 in 0 ..< self.size() {
             arr.append(self.getWith(i) as! T)
         }
         return arr
+    }
+}
+
+extension Array where Element : Any {
+    var toJavaArray : JavaUtilArrayList {
+        get {
+            let arrayList : JavaUtilArrayList = JavaUtilArrayList(int: jint(self.count))
+            self.forEach({ arrayList.add(withId: $0) })
+            return arrayList
+        }
     }
 }
